Add error handling to connect parameter
Previously, the parsing of the arguments to --connect would fall back to
client ID 0 in most failure cases due to use of atoi() to find potential
numeric arguments. Replacing that with strtol() and adding explicit
error checking to ensure that we only produce a numeric ID if the
argument is purely numeric, and otherwise fail with a useful message if
the client ID is not valid.

Catches error from snd_seq_connect_from() to exit with error if we fail
to open the requested MIDI client.

Signed-off-by: Brett T. Warden <4c0e8e88@tempmail.wgz.org>